adapter parameter becomes City list, dev-environment, cutting out test variables

+++ app/src/main/java/com/vvash17company/weatherapp/models/additional/HumidityData.kt
@@ -3,5 +3,5 @@ package com.vvash17company.weatherapp.models.additional
 data class HumidityData(
     var name: String,
     var iconId: String,
-    var percentage: Int
+    var percentage: Long
 )
